backend/requirements.txt
sentence-transformers==2.7.0
transformers==4.44.2
huggingface_hub==0.23.4
pdfplumber==0.11.4
pymongo==4.7.2
python-dotenv==1.0.1
